requirements.txt
agentforce-sdk
mcp
fastmcp
fastapi[standard]
sqlite3
python-dotenv
requests